webhook_server.py
from flask import Flask, request
from telegram import Update
from telegram.ext import Application, CommandHandler, CallbackQueryHandler, MessageHandler, filters, ContextTypes
import os
import logging
from telethon.sync import TelegramClient

# Настройка логирования
logging.basicConfig(
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
    level=logging.INFO
)
logger = logging.getLogger(__name__)

# Токен вашего бота
TOKEN = os.environ.get('7833414798:AAFpEUyslDz0TrzWupNC2nrAk9gY2nFWzio')

# Укажите ваши данные для Telethon
api_id = os.environ.get('29758240')
api_hash = os.environ.get('45aa1a0337bf2ab7c931f4fa6a45b344')
phone_number = os.environ.get('+380958153249')
session_file = 'user_session.session'  # Файл для сохранения сессии

# Инициализация клиента Telethon
telethon_client = TelegramClient(session_file, api_id, api_hash)

# Инициализация приложения Flask
app = Flask(__name__)

# Инициализация приложения Telegram
application = Application.builder().token(TOKEN).build()

# Маршрут для вебхуков
@app.route(f'/{TOKEN}', methods=['POST'])
def webhook():
    update = Update.de_json(request.get_json(force=True), application.bot)
    application.process_update(update)
    return 'ok'

# Функция для настройки бота
async def setup():
    await telethon_client.start(phone=phone_number)

def main():
    application.add_handler(CommandHandler('start', start))
    application.add_handler(CallbackQueryHandler(button_handler))
    application.add_handler(MessageHandler(filters.TEXT & (~filters.COMMAND), save_channel))

    # Запуск клиента Telethon
    asyncio.run(setup())

    # Запуск Flask-сервера
    app.run(port=8000, debug=True)

# Ваши функции бота
async def start(update: Update, context: ContextTypes.DEFAULT_TYPE):
    # Ваш код для обработки команды /start
    pass

async def button_handler(update: Update, context: ContextTypes.DEFAULT_TYPE):
    # Ваш код для обработки нажатий на кнопки
    pass

async def save_channel(update: Update, context: ContextTypes.DEFAULT_TYPE):
    # Ваш код для сохранения канала или группы
    pass

if __name__ == '__main__':
    main()